movement and render

+++ src/render_player_rotate.c
@@ -18,14 +18,10 @@ void player_rotate(t_var *var, int change)
     double change_angle;
     
     change_angle = (M_PI / 90) * change;
-    var->player.direction_x = var->player.direction_x * cos(change_angle) 
-    - var->player.direction_y * sin(change_angle);
-    var->player.direction_y = var->player.direction_x * sin(change_angle) 
-    + var->player.direction_y * cos(change_angle);
-    var->camerax = var->camerax * cos(change_angle) 
-    - var->cameray * sin(change_angle);
-    var->cameray = var->camerax * sin(change_angle) 
-    + var->cameray * cos(change_angle); 
+    var->player.direction_x = (int) var->player.direction_x * cos(change_angle) - var->player.direction_y * sin(change_angle);
+    var->player.direction_y = (int) var->player.direction_x * sin(change_angle) + var->player.direction_y * cos(change_angle);
+    var->camerax = var->camerax * cos(change_angle) - var->cameray * sin(change_angle);
+    var->cameray = var->camerax * sin(change_angle) + var->cameray * cos(change_angle); 
 }
 
 /*
